pyproject.toml
[project]
name = "mcpdev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "fastapi-mcp>=0.1.8",
    "google-generativeai>=0.8.4",
    "mcp[cli]",
    "pydantic>=2.11.3",
    "pydantic-ai>=0.0.55",
    "python-dotenv>=1.1.0",
    "uvicorn",
]
